Added function list spreadsheet
Spreadsheet is a list of all functions from both Martin's and Alastair's FemapAPI libraries.
First purpose is to identify double ups for merging the two libraries. Second purpose is to tag the functions with common functionalities such as action, object, and method. It will also note which functions are low level, e.g., "create_node", and which are more high level for specific applications, e.g., "create_nodal_stress". This should help to turn the libraries into a structured module as well as highlight any gaps in library.